+++ app/src/main/java/com/openclassrooms/realestatemanager/ui/addestate/AddEstateViewModel.kt
@@ -107,7 +107,7 @@ class AddEstateViewModel : ViewModel() {
             null,
             type.value!!,
             price.value!!.toInt(),
-            surface.value!!.toDouble(),
+            surface.value!!.toInt(),
             rooms.value!!.toInt(),
             bathrooms.value!!.toInt(),
             bedrooms.value!!.toInt(),
